Move `RunnerError` to `worker`
