--- packages/engine/src/hash_types/mod.rs
@@ -1,3 +1,0 @@
-pub mod worker;
-
-pub use self::error::{Error, Result};
