+++ packages/engine/src/lib.rs
@@ -41,7 +41,6 @@ pub mod env;
 mod error;
 pub mod experiment;
 pub mod fetch;
-pub mod hash_types;
 pub mod language;
 pub mod output;
 pub mod proto;
