+++ packages/engine/src/proto.rs
@@ -9,7 +9,7 @@ use serde_json::Value as SerdeValue;
 use stateful::globals::Globals;
 use uuid::Uuid;
 
-use crate::{hash_types::worker::RunnerError, simulation::status::SimStatus};
+use crate::simulation::status::SimStatus;
 
 // TODO: UNUSED: Needs triage
 pub type SerdeMap = serde_json::Map<String, SerdeValue>;
@@ -57,7 +57,10 @@ impl FromStr for ExperimentName {
 
 use crate::{
     simulation::enum_dispatch::enum_dispatch,
-    worker::runner::comms::outbound::{PackageError, UserError, UserWarning},
+    worker::{
+        runner::comms::outbound::{PackageError, UserError, UserWarning},
+        RunnerError,
+    },
 };
 
 /// The message type sent from the engine to the orchestrator.
