+++ packages/engine/src/simulation/controller/run.rs
@@ -7,7 +7,6 @@ use crate::{
     config::SimRunConfig,
     datastore::store::Store,
     experiment::controller::comms::{sim_status::SimStatusSend, simulation::SimCtlRecv},
-    hash_types::worker::RunnerError,
     output::SimulationOutputPersistenceRepr,
     proto::SimulationShortId,
     simulation::{
@@ -21,6 +20,7 @@ use crate::{
         package::run::Packages,
         status::SimStatus,
     },
+    worker::RunnerError,
 };
 
 enum LoopControl {
