+++ packages/engine/src/simulation/status.rs
@@ -1,10 +1,10 @@
 use serde::{Deserialize, Serialize};
 
 use crate::{
-    hash_types::worker::RunnerError,
     output::OutputPersistenceResultRepr,
     proto::SimulationShortId,
     simulation::{command::StopCommand, Result},
+    worker::RunnerError,
 };
 
 // Sent from sim runs to experiment main loop.
