+++ packages/engine/src/simulation/step_result.rs
@@ -1,7 +1,7 @@
 use crate::{
-    hash_types::worker::RunnerError,
     proto::SimulationShortId,
     simulation::{agent_control::AgentControl, step_output::SimulationStepOutput},
+    worker::RunnerError,
 };
 
 pub struct SimulationStepResult {
